♻️Refactor: infrastructure layer Package 분리  - 멘토링 피드백 반영   . jpa와 repository 파일 패키지 분리
