+++ src/main/java/kr/hhplus/be/infrastructure/coupon/repository/CouponPublishRepositoryImpl.java
@@ -1,9 +1,10 @@
-package kr.hhplus.be.infrastructure.coupon;
+package kr.hhplus.be.infrastructure.coupon.repository;
 
 import kr.hhplus.be.domain.coupon.dto.CouponSearchDTO;
 import kr.hhplus.be.domain.coupon.entity.CouponPublish;
 import kr.hhplus.be.domain.coupon.repository.CouponPublishRepository;
 import kr.hhplus.be.domain.user.dto.UserCouponDTO;
+import kr.hhplus.be.infrastructure.coupon.jpa.CouponPublishJpaRepository;
 import kr.hhplus.be.support.exception.BusinessException;
 import kr.hhplus.be.support.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
