+++ src/main/java/kr/hhplus/be/infrastructure/coupon/repository/CouponRepositoryImpl.java
@@ -1,14 +1,12 @@
-package kr.hhplus.be.infrastructure.coupon;
+package kr.hhplus.be.infrastructure.coupon.repository;
 
 import kr.hhplus.be.domain.coupon.entity.Coupon;
-import kr.hhplus.be.domain.user.dto.UserCouponDTO;
 import kr.hhplus.be.domain.coupon.repository.CouponRepository;
+import kr.hhplus.be.infrastructure.coupon.jpa.CouponJpaRepository;
 import kr.hhplus.be.support.exception.BusinessException;
 import kr.hhplus.be.support.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 @Repository
 @RequiredArgsConstructor
