+++ src/main/java/kr/hhplus/be/infrastructure/order/jpa/OrderJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.infrastructure.order;
+package kr.hhplus.be.infrastructure.order.jpa;
 
 import kr.hhplus.be.domain.order.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
