+++ src/main/java/kr/hhplus/be/infrastructure/order/repository/OrderDetailRepositoryImpl.java
@@ -1,7 +1,8 @@
-package kr.hhplus.be.infrastructure.order;
+package kr.hhplus.be.infrastructure.order.repository;
 
 import kr.hhplus.be.domain.order.entity.OrderDetail;
 import kr.hhplus.be.domain.order.repository.OrderDetailRepository;
+import kr.hhplus.be.infrastructure.order.jpa.OrderDetailJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
