+++ src/main/java/kr/hhplus/be/infrastructure/order/repository/OrderRepositoryImpl.java
@@ -1,7 +1,8 @@
-package kr.hhplus.be.infrastructure.order;
+package kr.hhplus.be.infrastructure.order.repository;
 
 import kr.hhplus.be.domain.order.entity.Order;
 import kr.hhplus.be.domain.order.repository.OrderRepository;
+import kr.hhplus.be.infrastructure.order.jpa.OrderJpaRepository;
 import kr.hhplus.be.support.exception.BusinessException;
 import kr.hhplus.be.support.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
