+++ src/main/java/kr/hhplus/be/infrastructure/payment/jpa/PaymentJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.infrastructure.payment;
+package kr.hhplus.be.infrastructure.payment.jpa;
 
 import kr.hhplus.be.domain.payment.entity.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
