+++ src/main/java/kr/hhplus/be/infrastructure/payment/repository/PaymentRepositoryImpl.java
@@ -1,7 +1,8 @@
-package kr.hhplus.be.infrastructure.payment;
+package kr.hhplus.be.infrastructure.payment.repository;
 
 import kr.hhplus.be.domain.payment.entity.Payment;
 import kr.hhplus.be.domain.payment.repository.PaymentRepository;
+import kr.hhplus.be.infrastructure.payment.jpa.PaymentJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
