+++ src/main/java/kr/hhplus/be/infrastructure/product/custom/ProductRepositoryCustom.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.infrastructure.product;
+package kr.hhplus.be.infrastructure.product.custom;
 
 import kr.hhplus.be.domain.product.dto.ProductSearchDTO;
 import kr.hhplus.be.domain.product.dto.TopSalesProductDTO;
