+++ src/main/java/kr/hhplus/be/infrastructure/product/jpa/ProductInventoryJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.infrastructure.product;
+package kr.hhplus.be.infrastructure.product.jpa;
 
 import kr.hhplus.be.domain.product.entity.ProductInventory;
 import org.springframework.data.jpa.repository.JpaRepository;
